Added an abstraction of the forms handling
